@@ -17,7 +17,7 @@ import org.springframework.http.HttpMethod;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.multipart.MultipartFile;
-import com.alibaba.fastjson.JSON;
+import com.alibaba.fastjson2.JSON;
 import com.oa.common.core.utils.ServletUtils;
 import com.oa.common.core.utils.StringUtils;
 import com.oa.common.core.utils.ip.IpUtils;
@@ -29,7 +29,7 @@ import com.oa.system.api.domain.SysOperLog;
 /**
  * 操作日志记录处理
  * 
- * @author ruoyi
+ * @author dragon
  */
 @Aspect
 @Component
