@@ -5,8 +5,8 @@ import com.google.code.kaptcha.text.impl.DefaultTextCreator;
 
 /**
  * 验证码文本生成器
- * 
- * @author ruoyi
+ *
+ * @author dragon
  */
 public class KaptchaTextCreator extends DefaultTextCreator
 {
@@ -20,7 +20,7 @@ public class KaptchaTextCreator extends DefaultTextCreator
         int x = random.nextInt(10);
         int y = random.nextInt(10);
         StringBuilder suChinese = new StringBuilder();
-        int randomoperands = (int) Math.round(Math.random() * 2);
+        int randomoperands = random.nextInt(3);
         if (randomoperands == 0)
         {
             result = x * y;
@@ -30,7 +30,7 @@ public class KaptchaTextCreator extends DefaultTextCreator
         }
         else if (randomoperands == 1)
         {
-            if (!(x == 0) && y % x == 0)
+            if ((x != 0) && y % x == 0)
             {
                 result = y / x;
                 suChinese.append(CNUMBERS[y]);
