@@ -1,8 +1,11 @@
 package com.oa.system.controller;
 
 import java.io.IOException;
+import java.util.Arrays;
 import java.util.Objects;
 
+import com.oa.common.core.utils.file.FileTypeUtils;
+import com.oa.common.core.utils.file.MimeTypeUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -124,9 +127,13 @@ public class SysProfileController extends BaseController
      */
     @Log(title = "用户头像", businessType = BusinessType.UPDATE)
     @PostMapping("/avatar")
-    public AjaxResult avatar(@RequestParam("avatarfile") MultipartFile file) throws IOException {
+    public AjaxResult avatar(@RequestParam("avatarfile") MultipartFile file) {
         if (!file.isEmpty()) {
             LoginUser loginUser = SecurityUtils.getLoginUser();
+            String extension = FileTypeUtils.getExtension(file);
+            if (!StringUtils.equalsAnyIgnoreCase(extension, MimeTypeUtils.IMAGE_EXTENSION)) {
+                return AjaxResult.error("文件格式不正确，请上传" + Arrays.toString(MimeTypeUtils.IMAGE_EXTENSION) + "格式");
+            }
             R<SysFileVo> fileResult = remoteFileService.upload(file, "avatar");
             if(Objects.equals(R.FAIL, fileResult.getCode())){
                 return AjaxResult.error("文件服务异常，请联系管理员");
